pyproject.toml
[tool.poetry]
name = "aio_send_dir"
version = "0.1.1"
description = "Gives you the ability to compress and send directory to a email"
authors = ["Aleksei Devyatkin <aleksei@devyatkin.dev>"]
readme = "README.md"
repository = "https://github.com/devalv/aio_send_dir"
packages = [{include = "aio_send_dir"}]

classifiers = [
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3 :: Only',
]

[tool.poetry.dependencies]
python = "^3.11"
aiofiles = {extras = ["os"], version = "^23.2.1"}
aiozipstream = "^0.4"
aiosmtplib = "^3.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
preview = true # (E271,E203) https://docs.astral.sh/ruff/preview/

unfixable = [
    "ERA", # do not autoremove commented out code
]

select = [
    "A", # same flake8-builtins
    "B", # same flake8-bugbear
    "E", # pycodestyle
    "F", # pyflakes
    "Q", # same flake8-quotes
    "W", # pycodestyle
    "E271", # Multiple spaces after keyword
    "ERA", # flake8-eradicate/eradicate
    "I", # same isort
    "T20",  # flake8-print
]

# Same as Black.(for future format in ruff)
line-length = 120

# Assume Python 3.11
target-version = "py311"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
